+++ R/availableCores.R
@@ -533,3 +533,102 @@ getNproc <- function(ignore = c("OMP_NUM_THREADS", "OMP_THREAD_LIMIT")) {
   
   NA_integer_
 }
+
+
+#' @section Protection against CPU overuse:
+#' Using too many parallel workers on the same machine may result in
+#' overusing the CPU.  For example, if an R script hard codes the
+#' number of parallel workers to 32, as in
+#'
+#' ```r
+#' cl <- makeClusterPSOCK(32)
+#' ```
+#'
+#' it will use more than 100% of the CPU cores when running on machine with
+#' fewer than 32 CPU cores.  For example, on a eight-core machine, this
+#' may run the CPU at 400% of its capacity, which has a significant
+#' negative effect on the current R process, but also on all other processes
+#' running on the same machine.  This also a problem on systems where R
+#' gets allotted a specific number of CPU cores, which is the case on
+#' high-performance compute (HPC) clusters, but also on other shared systems
+#' that limits user processes via Linux Control Groups (CGroups).
+#' For example, a free account on Posit Cloud is limited to a single
+#' CPU core. Parallelizing with 32 workers when only having access to
+#' a single core, will result in 3200% overuse and 32 concurrent R
+#' processes competing for this single CPU core.
+#'
+#' To protect against CPU overuse by mistake, `makeClusterPSOCK()` will
+#' warn when parallelizing above 100%;
+#'
+#' ```r
+#' cl <- parallelly:::makeClusterPSOCK(12, dryrun = TRUE)
+#' Warning message:
+#' In checkNumberOfLocalWorkers(workers) :
+#'   Careful, you are setting up 12 localhost parallel workers with
+#' only 8 CPU cores available for this R process, which could result
+#' in a 150% load. The maximum is set to 100%. Overusing the CPUs has
+#' negative impact on the current R process, but also on all other
+#' processes of yours and others running on the same machine. See
+#' help("parallelly.options", package = "parallelly") for how to
+#' override this threshold
+#' ```
+#'
+#' Any attempts resulting in more than 300% overuse will be refused;
+#'
+#' ```r
+#' > cl <- parallelly:::makeClusterPSOCK(25, dryrun = TRUE)
+#' Error in checkNumberOfLocalWorkers(workers) : 
+#'   Attempting to set up 25 localhost parallel workers with only
+#' 8 CPU cores available for this R process, which could result in
+#' a 312% load. The maximum is set to 300%. Overusing the CPUs has
+#' negative impact on the current R process, but also on all other
+#' processes of yours and others running on the same machine. See
+#' help("parallelly.options", package = "parallelly") for how to
+#' override this threshold
+#' ```
+#'
+#' See [parallelly.options] for how to change the default thresholds.
+#'
+#' @rdname makeClusterPSOCK
+checkNumberOfLocalWorkers <- function(workers) {
+  if (inherits(workers, "AsIs")) return()
+  
+  limits <- getOption("parallelly.maxWorkers.localhost", c(1.0, 3.0))
+  if (length(limits) == 0) return()
+
+  ## FIXME: Temporarily, ignore _R_CHECK_LIMIT_CORES_ limits
+  ## This will give a few packages time to be fixed. /HB 2024-02-09
+  ignore <- c("_R_CHECK_LIMIT_CORES_")
+  ignore <- getOption("parallelly.maxWorkers.localhost.ignore", ignore)
+  if (length(ignore) > 0) {
+    methods <- eval(formals(availableCores)$methods)
+    methods <- setdiff(methods, ignore)
+        ncores <- availableCores(methods = methods)
+  } else {
+    ncores <- availableCores()
+  }
+  reason <- names(ncores)
+  if (is.null(reason)) reason <- "N/A"
+  
+  rho <- workers / ncores
+
+  ## Produce an error?
+  if (length(limits) >= 2) {
+    if (rho > limits[2]) {
+      msg <- sprintf("Attempting to set up %d localhost parallel workers with only %d CPU cores available for this R process (per %s), which could result in a %.0f%% load", workers, ncores, sQuote(reason), 100 * workers / ncores)
+      msg <- sprintf("%s. The hard limit is set to %.0f%%", msg, 100 * limits[2])
+      msg <- sprintf("%s. Overusing the CPUs has negative impact on the current R process, but also on all other processes of yours and others running on the same machine", msg)
+      msg <- sprintf("%s. See help(\"parallelly.options\", package = \"parallelly\") for how to override the soft and hard limits", msg)
+      stop(msg)
+    }
+  }
+  
+  ## Warn?
+  if (rho > limits[1]) {
+    msg <- sprintf("Careful, you are setting up %d localhost parallel workers with only %d CPU cores available for this R process (per %s), which could result in a %.0f%% load", workers, ncores, sQuote(reason), 100 * workers / ncores)
+    msg <- sprintf("%s. The soft limit is set to %.0f%%", msg, 100 * limits[1])
+    msg <- sprintf("%s. Overusing the CPUs has negative impact on the current R process, but also on all other processes of yours and others running on the same machine", msg)
+    msg <- sprintf("%s. See help(\"parallelly.options\", package = \"parallelly\") for how to override the soft and hard limits", msg)
+    warning(msg)
+  }
+} ## checkNumberOfLocalWorkers()