+++ R/ports.R
@@ -17,11 +17,9 @@
 #' @return
 #' Returns an integer representing the first port among `ports` that
 #' can be opened.  If none can be opened, then `default` is returned.
-#' If port querying is not supported, as in R (< 4.0.0), then
-#' `default` is returned.
 #'
 #' @export
-freePort <- function(ports = 1024:65535, default = if (getRversion() >= "4.0.0") NA_integer_ else "random", randomize = TRUE) {
+freePort <- function(ports = 1024:65535, default = "random", randomize = TRUE) {
   if (is.character(default)) {
     default <- match.arg(default, choices = c("first", "random"))
   } else {
@@ -68,56 +66,14 @@ freePort <- function(ports = 1024:65535, default = if (getRversion() >= "4.0.0")
   ## Find first available port
   for (kk in seq_along(ports)) {
     port <- ports[kk]
-    free <- canPortBeUsed(port)
     
-    ## SPECIAL CASE: If it's not possible to query ports,
-    ## then use the first one
-    if (is.na(free)) return(default)
-
     ## Available?
-    if (free) return(port)
+    if (isTcpPortAvailable(port)) {
+      return(port)
+    }
   }
 
   default
-}
-
-
-#' Check whether a TCP port can be opened or not
-#'
-#' @param port (integer) A TCP port in \[0, 65535\].
-#'
-#' @return
-#' `canPortBeUsed(port)` returns a logical indicating whether the port can
-#' be opened or not, or cannot be queried.  If the port can be opened,
-#' then `TRUE` is returned, if cannot be opened then `FALSE` is returned,
-#' which may happen if the port is used by another process.
-#' If port querying is not supported, as in R (< 4.0.0),  then `NA` is
-#' returned.
-#'
-#' @keywords internal
-canPortBeUsed <- function(port) {
-  port <- assertPort(port)
-
-  ## If not possible to query, return NA
-  ## It works in R (>= 4.0.0)
-  ns <- asNamespace("parallel")
-  if (!exists("serverSocket", envir = ns, mode = "function")) return(NA)
-  serverSocket <- get("serverSocket", envir = ns, mode = "function")
-
-  ## suspendInterrupts() is available in R (>= 3.5.0), so we're good here,
-  ## but we use this to avoid 'R CMD check' WARNINGs in R (< 3.5.0)
-  suspendInterrupts <- get("suspendInterrupts", envir = asNamespace("base"), mode = "function")
-
-  ## Prevent user interrupts from giving false results
-  suspendInterrupts({
-    con <- tryCatch(serverSocket(port), error = identity)
-  })
-
-  ## Success?
-  free <- inherits(con, "connection")
-  if (free) close(con)
-  
-  free
 }
 
 
@@ -167,4 +123,102 @@ randomParallelPorts <- function(default = 11000:11999) {
   }
 
   from:to
+}
+
+
+#' Initialize R's internet module
+#'
+#' R needs to initialize its Internet module before we can create sockets.
+#' This is automatically done when R starts on Linux and macOS, but not on
+#' MS Windows.
+#'
+#' This function triggers the initialization, but calling one of R's
+#' built-in functions that does so.  See 'src/main/internet.c' in the
+#' R source code for where this happens under the hood.
+#'
+#' @noRd
+initialize_internet <- local({
+  done <- (.Platform[["OS.type"]] != "windows")
+  baseenv <- baseenv()
+  
+  function() {
+    ## Already done?
+    if (done) return()
+  
+    if (exists("serverSocket", mode = "function", envir = baseenv, inherits = FALSE)) {
+      ## R (>= 4.0.0)
+      serverSocket <- get("serverSocket", mode = "function", envir = baseenv, inherits = FALSE)
+      con <- serverSocket(port = 0L)
+      close(con)
+    } else {
+      ## R (< 4.0.0)
+      tryCatch({
+        con <- socketConnection(port = 0L, server = FALSE, blocking = FALSE, timeout = 0.0)
+        close(con)
+      }, error = identity)
+    }
+    
+    done <<- TRUE
+  }
+})
+
+
+#' Check whether a TCP port is available
+#'
+#' @param port (integer) TCP port in $\[1,65535\]$ to test.
+#'
+#' @param test One or more tests to apply.
+#' If `"bind"`, check if it is possible to _bind_ the TCP port.
+#' If `"listen"`, check if it is possible to _listen_ to the TCP port.
+#'
+#' @return
+#' Return TRUE if the TCP port is available, otherwise FALSE.
+#'
+#' @keywords internal
+#' @noRd
+isTcpPortAvailable <- function(port, test = c("bind", "listen")) {
+  stopifnot(
+    length(port) == 1L,
+    is.numeric(port),
+    !is.na(port),
+    port >= 1,
+    port <= 65535
+  )
+  port <- as.integer(port)
+  stopifnot(
+    port >= 1L,
+    port <= 65535L
+  )
+
+  test <- match.arg(test, several.ok = TRUE)
+
+  ## SPECIAL: Fake port availability?
+  if (nzchar(Sys.getenv("_R_PARALLELLY_CHECK_AVAILABLE_PORTS_"))) {
+    value <- Sys.getenv("_R_PARALLELLY_CHECK_AVAILABLE_PORTS_")
+    if (value == "any") {
+#      warning("parallelly:::isTcpPortAvailable() returns TRUE because _R_PARALLELLY_CHECK_AVAILABLE_PORTS_=any")
+      return(TRUE)
+    }
+    stop("Unknown value on _R_PARALLELLY_CHECK_AVAILABLE_PORTS_: ", sQuote(value))
+  }
+
+  initialize_internet()
+  res <- .Call(C_R_test_tcp_port, port)
+  
+  if (nzchar(Sys.getenv("R_PARALLELLY_DEBUG"))) {
+    reason <- if (res == 0) {
+      "available (can bind and listen)"
+    } else if (res %/% 10 == 1) {
+      "not available (cannot set up socket)"
+    } else if (res %/% 10 == 2) {
+      "not available (cannot reuse port in TIME_WAIT state)"
+    } else if (res %/% 10 == 3) {
+      "not available (cannot bind to port)"
+    } else if (res %/% 10 == 4) {
+      "not available (cannot listen)"
+    }
+    message(sprintf("parallelly:::isTcpPortAvailable(%d): %s", port, reason))
+  }
+  
+  (res == 0L)
 }