+++ src/000.api.h
@@ -1,0 +1,2 @@
+/* C-level API that is called from R */
+SEXP R_test_tcp_port(SEXP port);