+++ NEWS.md
@@ -1,3 +1,35 @@
+# Version 1.37.0 [2024-02-14]
+
+## New Features
+
+ * `makeClusterPSOCK(nworkers)` gained protection against setting up
+   too many localhost workers relative to number of available CPU
+   cores. If `nworkers / availableCores()` is greater than 1.0 (100%),
+   then a warning is produced. If greater than 3.0 (300%), an error is
+   produced.  These limits can be configured by R option
+   `parallelly.maxWorkers.localhost`. These checks are skipped if
+   `nworkers` inherits from `AsIs`, e.g. `makeClusterPSOCK(I(16))`.
+   The current 3.0 (300%) limit is likely to be decreased in a future
+   release.  A few packages fail `R CMD check --as-cran` with this
+   validation enabled. For example, one package uses 8 parallel
+   workers in its examples, while `R CMD check --as-cran` only allows
+   for two. To give such packages time to be fixed, the CRAN-enforced
+   limits are ignored for now.
+ 
+## Miscellaneous
+
+ * `makeClusterPSOCK()` could produce a confusing error `Invalid port:
+   NA` if a non-available port was requested.  Now the error message
+   is more informative, e.g.
+   `Argument 'port' specifies non-available port(s): 80`.
+
+## Bug Fixes
+
+ * The internal method for checking if a TCP port is available has
+   been improved.  Previously, it could incorrectly conclude a port
+   was available, when it was not.
+
+
 # Version 1.36.0 [2023-05-26]
 
 ## New Features